requirements.txt
msgpack == 1.*
pywebview == 5.*
